requirements.txt
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
dj-rest-auth==7.0.1
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
executing==2.2.1
idna==3.10
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.4
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
